+++ bundles/org.corpus_tools.hexatomic.formats/src/main/java/org/corpus_tools/hexatomic/formats/importer/ImporterSelectionPage.java
@@ -60,6 +60,7 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
     btnExb.setSelection(false);
     btnPaulaXml.setSelection(false);
     btnGraphAnno.setSelection(false);
+    btnTxt.setSelection(false);
     Optional<Pepper> pepper = Activator.getPepper();
     setPageComplete(false);
 
@@ -81,6 +82,8 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
             setPageComplete(true);
           } else if (format.get() == ImportFormat.GRAPHANNO) {
             btnGraphAnno.setSelection(true);
+          } else if (format.get() == ImportFormat.TXT) {
+            btnTxt.setSelection(true);
             setPageComplete(true);
           }
         } else {
@@ -103,6 +106,8 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
       return Optional.of(ImportFormat.PAULA);
     } else if (btnGraphAnno.getSelection()) {
       return Optional.of(ImportFormat.GRAPHANNO);
+    } else if (btnTxt.getSelection()) {
+      return Optional.of(ImportFormat.TXT);
     }
     return Optional.empty();
   }