+++ CHANGELOG.md
@@ -8,7 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add import of plain text files with optional tokenization (#110)
 - Add the GraphAnno format to the supported import formats (#112)
+
+### Fixed
+
+- Fixed some minor typos in the GUI
+
+## [0.7.1] - 2021-10-26
+
+### Fixed
+
+- Remove and ignore .settings folders from repository to enable smoother import into Eclipse (#332)
+- Fix partially failing integration test by determining file menu programmatically rather than relying on defaults
 
 ## [0.7.0] - 2021-09-28
 
@@ -189,7 +201,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2019-11-01
 
 - This is the first public preview release of the Hexatomic developer platform.
-[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/hexatomic/hexatomic/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/hexatomic/hexatomic/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/hexatomic/hexatomic/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/hexatomic/hexatomic/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/hexatomic/hexatomic/compare/v0.5.1...v0.6.0